force use of ipv4 only for mia
this is caused by the fact, that mia got only IPv6 address.

When ssh is forced to use IPv4 only (like in a IPv4-only network
as SAP Internet) there is no problem - mia resolves to a IPv4 and v6,
but a connection is only possible via v4. And this IPv4 interface is
the management interface where tcp/443 is a sshd.

So for IPv6 we need to use another (unavailable) IP. As this is currently
not available I forced the ssh-client to use IPv4. (this is somehow aaaaaah).